+++ web/modules/medulla_server/includes/xmlrpc.inc.php
@@ -1,16 +1,18 @@
 <?php
-/**
+/*
  * (c) 2004-2007 Linbox / Free&ALter Soft, http://linbox.com
- * (c) 2007-2008 Mandriva, http://www.mandriva.com/
+ * (c) 2007 Mandriva, http://www.mandriva.com
+ * (c) 2016-2023 Siveo, http://www.siveo.net
+ * (c) 2024-2025 Medulla, http://www.medulla-tech.io
  *
  * $Id$
  *
- * This file is part of Mandriva Management Console (MMC).
+ * This file is part of MMC, http://www.medulla-tech.io
  *
  * MMC is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
+ * the Free Software Foundation; either version 3 of the License, or
+ * any later version.
  *
  * MMC is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -18,10 +20,9 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with MMC; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ * along with MMC; If not, see <http://www.gnu.org/licenses/>.
+ *
  */
-
 function displayLocalisationBar() {
     if (!isset($_SESSION["displayLocalisationBar"])) {
         $_SESSION["displayLocalisationBar"] = xmlCall("medulla_server.displayLocalisationBar");
@@ -60,5 +61,13 @@ function xmlrpc_setfrommedulla_serverlogxmpp(   $text,
                                                     $action,
                                                     $touser,
                                                     $fromuser));
+}
+
+function xmlrpc_restart_all_services(){
+    return xmlCall("admin.restart_medulla_services", array());
+}
+
+function xmlrpc_regenerate_agent(){
+    return xmlCall("admin.regenerate_agent", array());
 }
 ?>